- Kotlin Conversion Complete

+++ src/main/kotlin/wowsft/model/gameparams/ship/component/engine/Engine.kt
@@ -6,16 +6,16 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 @WoWSFT
 @JsonIgnoreProperties(ignoreUnknown = true)
 class Engine {
-    var backwardEngineForsag = 0.toDouble()
-    var backwardEngineForsagMaxSpeed = 0.toDouble()
-    var backwardEngineUpTime = 0.toDouble()
-    var backwardSpeedOnFlood = 0.toDouble()
-    var damagedEnginePowerMultiplier = 0.toDouble()
-    var damagedEnginePowerTimeMultiplier = 0.toDouble()
-    var forwardEngineForsag = 0.toDouble()
-    var forwardEngineForsagMaxSpeed = 0.toDouble()
-    var forwardEngineUpTime = 0.toDouble()
-    var forwardSpeedOnFlood = 0.toDouble()
-    var histEnginePower = 0.toDouble()
-    var speedCoef = 0.toDouble()
+    var backwardEngineForsag = 0.0
+    var backwardEngineForsagMaxSpeed = 0.0
+    var backwardEngineUpTime = 0.0
+    var backwardSpeedOnFlood = 0.0
+    var damagedEnginePowerMultiplier = 0.0
+    var damagedEnginePowerTimeMultiplier = 0.0
+    var forwardEngineForsag = 0.0
+    var forwardEngineForsagMaxSpeed = 0.0
+    var forwardEngineUpTime = 0.0
+    var forwardSpeedOnFlood = 0.0
+    var histEnginePower = 0.0
+    var speedCoef = 0.0
 }
