+++ src/main/kotlin/wowsft/model/gameparams/ship/component/firecontrol/FireControl.kt
@@ -4,6 +4,6 @@ import wowsft.config.WoWSFT
 
 @WoWSFT
 class FireControl {
-    var maxDistCoef = 0.toDouble()
-    var sigmaCountCoef = 0.toDouble()
+    var maxDistCoef = 0.0
+    var sigmaCountCoef = 0.0
 }
